+++ examples/demo-particles/src/commonMain/kotlin/com/pandulapeter/kubriko/demoParticles/implementation/actors/DemoParticleState.kt
@@ -51,7 +51,7 @@ internal class DemoParticleState(
         body.scale = Scale.Unit
     }
 
-    override fun update(deltaTimeInMilliseconds: Float): Boolean {
+    override fun update(deltaTimeInMilliseconds: Int): Boolean {
         currentProgress = 1f - (remainingLifespan / lifespanInMilliseconds)
         if (currentProgress >= 1) {
             return false
